requirements.txt
streamlit
pdf2image
paddleocr
opencv-python
numpy
langchain
sentence-transformers
faiss-cpu
openai